MS HM_Associates - integration stage 3

+++ Microservice/JAVA/Associates - Manager/employee/src/main/java/com/example/ctsproject/controller/EmployeeController.java
@@ -3,46 +3,89 @@ package com.example.ctsproject.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.ctsproject.entity.Employee;
-import com.example.ctsproject.exception.EmployeeEmptyException;
-import com.example.ctsproject.exception.EmployeeNotFoundException;
 import com.example.ctsproject.service.EmployeeService;
+import com.example.ctsproject.client.AuthClient;
+import com.example.ctsproject.entity.AuthResponse;
+import com.example.ctsproject.exception.AuthorizationException;
 
 @RestController
 public class EmployeeController {
 	@Autowired
 	EmployeeService employeeService;
+	@Autowired
+	AuthClient authClient;
+
 	
-	@GetMapping(value="/getemployee/{id}")
-	public Employee getEmployeeById(@PathVariable("id") int id) throws EmployeeNotFoundException{
-		return employeeService.getEmployeeById(id);
+	public boolean checkAuthentication(String requestTokenHeader) throws Exception {
+		ResponseEntity<AuthResponse> valid = authClient.getValidity(requestTokenHeader);
+		return valid.getBody().isValid();
+	}
+		
+	@GetMapping("/getemployee/{id}")
+	public Employee getEmployeeById(@PathVariable("id") int id,
+			@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
+
+		if (checkAuthentication(requestTokenHeader)) {
+			return employeeService.getEmployeeById(id);
+		} else {
+			throw new AuthorizationException("Invalid token");
+		}
 	}
 	
-	@PostMapping(value="/addemployee")
-	public Employee addEmployee(@RequestBody Employee employee){
-		return employeeService.addEmployee(employee);
+	@PostMapping("/addemployee")
+	public Employee addEmployee(@RequestBody Employee employee,
+			@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
+
+		if (checkAuthentication(requestTokenHeader)) {
+			return employeeService.addEmployee(employee);
+		} else {
+			throw new AuthorizationException("Invalid token");
+		}
 	}
 	
-	@GetMapping(value="/getemployee")
-	public List<Employee> getEmployee() throws EmployeeEmptyException{
-		return employeeService.getEmployee();
+	@GetMapping("/getemployee")
+	public List<Employee> getEmployee(
+			@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
+
+		if (checkAuthentication(requestTokenHeader)) {
+			return employeeService.getEmployee();
+		} else {
+			throw new AuthorizationException("Invalid token");
+		}
+
 	}
 	
 	@PutMapping("/updateemp/{id}")
-	public Employee updateEmployee(@PathVariable int id, @RequestBody Employee employee) {
-		return employeeService.updateEmployee(id, employee);
+	public Employee updateEmployee(@PathVariable("id") int id, @RequestBody Employee employee,
+			@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
+
+		if (checkAuthentication(requestTokenHeader)) {
+			return employeeService.updateEmployee(id, employee);
+		} else {
+			throw new AuthorizationException("Invalid token");
+		}
 	}
 	
-	@PutMapping("/deleteemp/{id}")
-	public Employee deleteEmployee(@PathVariable int id, @RequestBody boolean isDeleted) throws Exception {
-		return employeeService.deleteEmployee(id, isDeleted);
-	}
 	
+	@DeleteMapping("/deleteemp/{id}")
+	public Employee deleteEmployee(@PathVariable("id") int id, @RequestBody boolean isDeleted,
+			@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
+
+		if (checkAuthentication(requestTokenHeader)) {
+			return employeeService.deleteEmployee(id, isDeleted);
+		} else {
+			throw new AuthorizationException("Invalid token");
+		}
+	}
 }
